Rewrite CanvasViewport to use float coordinates and smooth animations

+++ Source/Pd/Patch.h
@@ -134,7 +134,7 @@ public:
     int splitViewIndex = 0;
     int windowIndex = 0;
 
-    Point<int> lastViewportPosition = { 1, 1 };
+    Point<float> lastViewportPosition = { 1, 1 };
     float lastViewportScale;
 
     SmallString lastUndoSequence;
